# Add deferred decoding support for composite values