--- a/runtime/interpreter/decode.go
+++ b/runtime/interpreter/decode.go
@@ -615,12 +615,13 @@ func (d *DecoderV4) decodeComposite(path []string) (*CompositeValue, error) {
 
 	if err != nil {
 		if e, ok := err.(*cbor.WrongTypeError); ok {
-			panic(fmt.Errorf("invalid composite encoding (@ %s): %s",
+			return nil, fmt.Errorf(
+				"invalid composite encoding (@ %s): %s",
 				strings.Join(path, "."),
 				e.ActualType.String(),
-			))
+			)
 		}
-		panic(err)
+		return nil, err
 	}
 
 	compositeValue := NewDeferredCompositeValue(path, content, d.owner)
@@ -1624,4 +1625,188 @@ func (d *DecoderV4) decodeCapabilityStaticType() (StaticType, error) {
 	return CapabilityStaticType{
 		BorrowType: borrowStaticType,
 	}, nil
+}
+
+// decodeCompositeMetaInfo decodes the meta info from the byte content and updates the composite value.
+// Meta info includes:
+//    - location
+//    - qualifiedIdentifier
+//    - kind
+//
+// This also extracts out the fields raw content and cache it separately inside the value.
+//
+func decodeCompositeMetaInfo(v *CompositeValue, content []byte) error {
+	d, err := NewByteDecoder(content, v.Owner, CurrentEncodingVersion, nil)
+	if err != nil {
+		return err
+	}
+
+	const expectedLength = encodedCompositeValueLength
+
+	size, err := d.decoder.DecodeArrayHead()
+
+	if err != nil {
+		if e, ok := err.(*cbor.WrongTypeError); ok {
+			return fmt.Errorf("invalid composite encoding (@ %s): expected [%d]interface{}, got %s",
+				strings.Join(v.decodePath, "."),
+				expectedLength,
+				e.ActualType.String(),
+			)
+		}
+		return err
+	}
+
+	if size != expectedLength {
+		return fmt.Errorf("invalid composite encoding (@ %s): expected [%d]interface{}, got [%d]interface{}",
+			strings.Join(v.decodePath, "."),
+			expectedLength,
+			size,
+		)
+	}
+
+	// Location
+
+	// Decode location at array index encodedCompositeValueLocationFieldKey
+	location, err := d.decodeLocation()
+	if err != nil {
+		return fmt.Errorf(
+			"invalid composite location encoding (@ %s): %w",
+			strings.Join(v.decodePath, "."),
+			err,
+		)
+	}
+
+	// Skip obsolete element at array index encodedCompositeValueTypeIDFieldKey
+	err = d.decoder.Skip()
+	if err != nil {
+		return err
+	}
+
+	// Kind
+
+	// Decode kind at array index encodedCompositeValueKindFieldKey
+	encodedKind, err := d.decoder.DecodeUint64()
+	if err != nil {
+		if e, ok := err.(*cbor.WrongTypeError); ok {
+			return fmt.Errorf(
+				"invalid composite kind encoding (@ %s): %s",
+				strings.Join(v.decodePath, "."),
+				e.ActualType.String(),
+			)
+		}
+		return err
+	}
+
+	kind := common.CompositeKind(encodedKind)
+
+	// Fields
+
+	fieldsContent, err := d.decoder.DecodeRawBytes()
+	if err != nil {
+		if e, ok := err.(*cbor.WrongTypeError); ok {
+			return fmt.Errorf(
+				"invalid composite fields encoding (@ %s): %s",
+				strings.Join(v.decodePath, "."),
+				e.ActualType.String(),
+			)
+		}
+		return err
+	}
+
+	v.fieldsContent = fieldsContent
+
+	// Qualified identifier
+
+	// Decode qualified identifier at array index encodedCompositeValueQualifiedIdentifierFieldKey
+	qualifiedIdentifier, err := d.decoder.DecodeString()
+	if err != nil {
+		if e, ok := err.(*cbor.WrongTypeError); ok {
+			return fmt.Errorf(
+				"invalid composite qualified identifier encoding (@ %s): %s",
+				strings.Join(v.decodePath, "."),
+				e.ActualType.String(),
+			)
+		}
+		return err
+	}
+
+	v.location = location
+	v.qualifiedIdentifier = qualifiedIdentifier
+	v.kind = kind
+
+	return nil
+}
+
+// decodeCompositeFields decodes fields from the byte content and updates the composite value.
+//
+func decodeCompositeFields(v *CompositeValue, content []byte) error {
+	d, err := NewByteDecoder(content, v.Owner, CurrentEncodingVersion, nil)
+	if err != nil {
+		return err
+	}
+
+	// Decode fields at array index encodedCompositeValueFieldsFieldKey
+	fieldsSize, err := d.decoder.DecodeArrayHead()
+	if err != nil {
+		if e, ok := err.(*cbor.WrongTypeError); ok {
+			return fmt.Errorf(
+				"invalid composite fields encoding (@ %s): %s",
+				strings.Join(v.decodePath, "."),
+				e.ActualType.String(),
+			)
+		}
+		return err
+	}
+
+	if fieldsSize%2 == 1 {
+		return fmt.Errorf(
+			"invalid composite fields encoding (@ %s): fields should have even number of elements: got %d",
+			strings.Join(v.decodePath, "."),
+			fieldsSize,
+		)
+	}
+
+	fields := NewStringValueOrderedMap()
+
+	// Pre-allocate and reuse valuePath.
+	valuePath := append(v.decodePath, "")
+
+	lastValuePathIndex := len(v.decodePath)
+
+	for i := 0; i < int(fieldsSize); i += 2 {
+
+		// field name
+		fieldName, err := d.decoder.DecodeString()
+		if err != nil {
+			if e, ok := err.(*cbor.WrongTypeError); ok {
+				return fmt.Errorf(
+					"invalid composite field name encoding (@ %s, %d): %s",
+					strings.Join(v.decodePath, "."),
+					i/2,
+					e.ActualType.String(),
+				)
+			}
+			return err
+		}
+
+		// field value
+
+		valuePath[lastValuePathIndex] = fieldName
+
+		decodedValue, err := d.decodeValue(valuePath)
+		if err != nil {
+			return fmt.Errorf(
+				"invalid composite field value encoding (@ %s, %s): %w",
+				strings.Join(v.decodePath, "."),
+				fieldName,
+				err,
+			)
+		}
+
+		fields.Set(fieldName, decodedValue)
+	}
+
+	v.fields = fields
+
+	return nil
 }