--- a/runtime/interpreter/encode.go
+++ b/runtime/interpreter/encode.go
@@ -991,7 +991,7 @@ func (e *Encoder) encodeCompositeValue(
 	}
 
 	// Encode location at array index encodedCompositeValueLocationFieldKey
-	err = e.encodeLocation(v.Location)
+	err = e.encodeLocation(v.location)
 	if err != nil {
 		return err
 	}