--- a/runtime/interpreter/encode.go
+++ b/runtime/interpreter/encode.go
@@ -1009,7 +1009,8 @@ func (e *Encoder) encodeCompositeValue(
 	}
 
 	// Encode fields (as array) at array index encodedCompositeValueFieldsFieldKey
-	err = e.enc.EncodeArrayHead(uint64(v.Fields.Len() * 2))
+	fields := v.Fields()
+	err = e.enc.EncodeArrayHead(uint64(fields.Len() * 2))
 	if err != nil {
 		return err
 	}
@@ -1020,7 +1021,7 @@ func (e *Encoder) encodeCompositeValue(
 
 	lastValuePathIndex := len(path)
 
-	for pair := v.Fields.Oldest(); pair != nil; pair = pair.Next() {
+	for pair := fields.Oldest(); pair != nil; pair = pair.Next() {
 		fieldName := pair.Key
 
 		// Encode field name as fields array element