--- a/runtime/interpreter/interpreter.go
+++ b/runtime/interpreter/interpreter.go
@@ -1308,7 +1308,7 @@ func (interpreter *Interpreter) declareNonEnumCompositeValue(
 			value := &CompositeValue{
 				location:            location,
 				qualifiedIdentifier: qualifiedIdentifier,
-				Kind:                declaration.CompositeKind,
+				kind:                declaration.CompositeKind,
 				fields:              fields,
 				InjectedFields:      injectedFields,
 				Functions:           functions,
@@ -1404,7 +1404,7 @@ func (interpreter *Interpreter) declareEnumConstructor(
 		caseValue := &CompositeValue{
 			location:            location,
 			qualifiedIdentifier: qualifiedIdentifier,
-			Kind:                declaration.CompositeKind,
+			kind:                declaration.CompositeKind,
 			fields:              caseValueFields,
 			// NOTE: new value has no owner
 			Owner:    nil,