Closes #845

## Description

Composite values have four storable properties:
- location
- qualified-Identifier
- kind
- fields

When reading a value from storage, only these four properties are decoded. Once the decoding is delayed, these properties may not have been properly initialized for a value returned from the decoder. Hence accessing them directly is unsafe.

This PR introduces a getter for each of the four properties. The getter function ensures the property is decoded and properly loaded.

### Approach:
- Composite value has two new extra properties:
  - `content` - Holds the raw content. Contains meta-info +  fields-content.
  - `fieldsContent` - Initially `nil`. Gets populated when the meta info is loaded. Hold the raw byte content related to composite value's fields
- When one of the meta info fields is accessed, all the meta info is loaded. At the same time, fields-content is also extracted out from the `content` and cached in `fieldsContent`.
- One drawback is accessing a field will also load the meta-info. This is a limitation in the storage format.

### Overhead
- There are two overheads:
  - Now we have to make a function call instead of directly using a field.
  - Function call always does a nil check.
    - Alternative is to store the getter as a local-field (function pointer) in the composite value, and replace it with a direct-field-returning function after the first invocation.
    - But the benchmarking showed that invoking a nil-checking static function is 5x faster than invoking a function pointer.
- However, both of these turned out to be very **trivial** at `go` level. It doesn't impact the overall performance of the runtime.

### Performance
For a composite value of `Foo` type, where :
```kotlin
struct Foo {
  var fname:   String
  var lname:   String
  var age:     Int
  var status:  String
  var address: Address
} 

struct Address {
  var street:  String
  var city:    String
  var state:   String
  var country: String
} 
```
Decoding perf numbers are:
```
name                                            old time/op    new time/op    delta
CompositeDeferredDecoding/Simply_decode-12        5.33µs ± 6%    1.45µs ± 5%  -72.86%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Access_identifier-12    5.69µs ± 7%    2.23µs ±15%  -60.87%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Access_field-12         5.91µs ± 2%    5.44µs ± 4%   -7.91%  (p=0.016 n=4+5)
CompositeDeferredDecoding/Re-encode_decoded-12    2.43µs ±26%    1.76µs ±13%  -27.68%  (p=0.008 n=5+5)

name                                            old alloc/op   new alloc/op   delta
CompositeDeferredDecoding/Simply_decode-12        2.61kB ± 0%    0.55kB ± 0%  -78.83%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Access_identifier-12    2.61kB ± 0%    0.86kB ± 0%  -66.87%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Access_field-12         2.82kB ± 0%    2.42kB ± 0%  -14.20%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Re-encode_decoded-12    1.16kB ± 0%    0.96kB ± 0%  -17.24%  (p=0.008 n=5+5)

name                                            old allocs/op  new allocs/op  delta
CompositeDeferredDecoding/Simply_decode-12          67.0 ± 0%       6.0 ± 0%  -91.04%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Access_identifier-12      67.0 ± 0%      13.0 ± 0%  -80.60%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Access_field-12           69.0 ± 0%      48.0 ± 0%  -30.43%  (p=0.008 n=5+5)
CompositeDeferredDecoding/Re-encode_decoded-12      14.0 ± 0%      10.0 ± 0%  -28.57%  (p=0.008 n=5+5)
```
_*'Simply_decode' - Only decode the value. Doesn't access the content of the value, like fields, type, etc._
______

For contributor use:

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels